test commit

+++ lib/binaries.sh
@@ -15,8 +15,8 @@ install_nodejs() {
   if needs_resolution "$requested_version"; then
     BP_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && cd .. && pwd )"
     versions_as_json=$(ruby -e "require 'yaml'; print YAML.load_file('$BP_DIR/manifest.yml')['dependencies'].select {|dep| dep['name'] == 'node' }.map {|dep| dep['version']}")
-    stable_major_version=4
-    resolved_version=$($BP_DIR/bin/node $BP_DIR/lib/version_resolver.js "$requested_version" "$versions_as_json" "$stable_major_version")
+    default_version=$($BP_DIR/compile-extensions/bin/default_version_for $BP_DIR/manifest.yml node)
+    resolved_version=$(ruby $BP_DIR/lib/version_resolver.rb "$requested_version" "$versions_as_json" "$default_version")
   fi
 
   if [[ "$resolved_version" = "undefined" ]]; then
@@ -24,9 +24,21 @@ install_nodejs() {
   else
     echo "Downloading and installing node $resolved_version..."
   fi
-  local download_url="https://s3pository.heroku.com/node/v$resolved_version/node-v$resolved_version-$os-$cpu.tar.gz"
-  curl "`translate_dependency_url $download_url`" --silent --fail --retry 5 --retry-max-time 15 -o /tmp/node.tar.gz || (>&2 $BP_DIR/compile-extensions/bin/recommend_dependency $download_url && false)
-  echo "Downloaded [`translate_dependency_url $download_url`]"
+
+  local heroku_url="https://s3pository.heroku.com/node/v$resolved_version/node-v$resolved_version-$os-$cpu.tar.gz"
+  local exit_code=0
+  local filtered_url=""
+
+  filtered_url=$($BP_DIR/compile-extensions/bin/download_dependency $heroku_url /tmp) || exit_code=$?
+  if [ $exit_code -ne 0 ]; then
+    echo -e "`$BP_DIR/compile-extensions/bin/recommend_dependency $heroku_url`" 1>&2
+    exit 22
+  fi
+
+  local downloaded_file=$(ls /tmp/node-v*.tar.gz)
+  mv $downloaded_file /tmp/node.tar.gz
+
+  echo "Downloaded [$filtered_url]"
   tar xzf /tmp/node.tar.gz -C /tmp
   rm -rf $dir/*
   mv /tmp/node-v$resolved_version-$os-$cpu/* $dir
